From me

+++ backend/init_db.py
@@ -59,7 +59,7 @@ def create_database_and_tables():
             id INT AUTO_INCREMENT PRIMARY KEY,
             restaurant_id INT NOT NULL,
             name VARCHAR(255) NOT NULL,
-            price DECIMAL(10, 2) NOT NULL,
+            price INT NOT NULL,
             description TEXT,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             FOREIGN KEY (restaurant_id) REFERENCES restaurants(id) ON DELETE CASCADE
@@ -79,8 +79,8 @@ def create_database_and_tables():
             customer_phone VARCHAR(20) NOT NULL,
             delivery_address VARCHAR(255) NOT NULL,
             payment_method VARCHAR(50) NOT NULL,
-            total_price DECIMAL(10, 2) NOT NULL,
-            tax DECIMAL(10, 2) NOT NULL,
+            total_price INT NOT NULL,
+            tax INT NOT NULL,
             status ENUM('on_process', 'on_delivery', 'delivered') DEFAULT 'on_process',
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
@@ -97,7 +97,7 @@ def create_database_and_tables():
             order_id INT NOT NULL,
             menu_id INT NOT NULL,
             quantity INT NOT NULL,
-            price DECIMAL(10, 2) NOT NULL,
+            price INT NOT NULL,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             FOREIGN KEY (order_id) REFERENCES orders(id) ON DELETE CASCADE,
             FOREIGN KEY (menu_id) REFERENCES menu_items(id) ON DELETE CASCADE
